msm: adsprpc: Fix to avoid untrusted pointer dereference
Currently, the compat ioctl call distinguishes itself using a global
flag. If a user sends a compat ioctl call followed by a normal ioctl
call, it may result in using a user passed address as a kernel address
in the fastrpcdriver. To address this issue, consider localizing the
compat flag for the ioctl call.

Acked-by: Ramesh Nallagopu <rnallago@qti.qualcomm.com>
Change-Id: I73923378c747efaa3beb6ccad460c15c7b63edcf
Signed-off-by: Santosh Sakore <quic_ssakore@quicinc.com>