smblite: Update EXTCON_CHG_USB_FAST in shim instead of in smblite-lib
EXTCON_CHG_USB_FAST is our addition to QCOM's existing extcon device
and is used by modules that operate on extcons to know when power was
connected, as an alternative to getting a reference to the power supply
and listening for power supply changed events.

Therefore it is important to keep this state in sync with what the
shim reports for the power supply present property.

To achieve this:
- Move extcon state update for EXTCON_CHG_USB_FAST to smblite-shim.
- Update this extcon state using a work queue. When either the usbin
  plugin status or the fake present votable vote changes, schedule a
  work item on the global workqueue to update the extcon state.
- Report that the extcon is connected if either the usb is plugged in,
  or, if the fake present votable has an active vote. Otherwise report
  disconnected.
- Only send an extcon update if the state changes from connected to
  disconnected or vice versa. Since extcon_set_state_sync() always
  triggers registered notifiers, this is just an optimization.

Also, enable dynamic debug for the smblite-shim, so that we can use
pr_debug() statements in this source file.

Bug: 355475618
Change-Id: I2f70fe83adab02a6a6cb9c0c33d9fd1f7f134d2e
Signed-off-by: Andrei Ciubotariu <aciubotariu@google.com>