soc: qcom: ramdump: Abort user-space read if timed-out
It's possible that the user-space process can be woken up
very late to collect the ramdumps by which time, the ramdump
collection timer would have expired. Following this, the PIL
driver unmaps the memory and continues. The user-space process
has no way to know about this and can make an attempt to read
the unmapped memory, thus resulting in a data abort.

Hence, notify the user-space process that the ramdump timer
has expired such that it can reset its state machine.

Also add the srcu related symbols to the allowed-list and update
the ABI snapshot accordingly.

Change-Id: I64414dcc37b450f4b90829b95ce6aa9347005c3c
Signed-off-by: Sudeepgoud Patil <sudeepgo@qti.qualcomm.com>